Merge to main

+++ codeine_django/community/views_article_comments.py
@@ -7,16 +7,16 @@ from rest_framework.decorators import api_view, permission_classes, parser_class
 from rest_framework.response import Response
 
 from rest_framework.permissions import (
-    IsAuthenticated,
-    IsAdminUser,
+    IsAuthenticatedOrReadOnly,
 )
-from .models import Article, ArticleComment
-from .serializers import ArticleCommentSerializer
-from common.models import Member
+from .models import Article, ArticleComment, ArticleCommentEngagement
+from .serializers import NestedArticleCommentSerializer
 
 # Create your views here.
+
+
 @api_view(['GET', 'POST'])
-@permission_classes((IsAuthenticated,))
+@permission_classes((IsAuthenticatedOrReadOnly,))
 def article_comment_view(request, article_id):
     '''
     Retrieves all article comments
@@ -24,17 +24,23 @@ def article_comment_view(request, article_id):
     if request.method == 'GET':
         article = Article.objects.get(pk=article_id)
         article_comments = ArticleComment.objects.filter(article=article)
-        
+
+        # return first two levels of comments under article
+        article_comments = article_comments.filter(reply_to=None)
+
         # extract query params
         search = request.query_params.get('search', None)
 
         if search is not None:
             article_comments = article_comments.filter(
+                Q(user__first_name__icontains=search) |
+                Q(user__last_name__icontains=search) |
                 Q(comment__icontains=search)
             )
         # end if
 
-        serializer = ArticleCommentSerializer(article_comments.all(), many=True)
+        serializer = NestedArticleCommentSerializer(
+            article_comments.all(), many=True, context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
     # end if
 
@@ -44,21 +50,28 @@ def article_comment_view(request, article_id):
     if request.method == 'POST':
         user = request.user
         data = request.data
-        article = Article.objects.get(pk=article_id)
-        parent_comment = None
-        if 'parent_comment_id' in data:
-            parent_comment = ArticleComment.objects.get(pk=data['parent_comment_id'])
 
         try:
+            article = Article.objects.get(pk=article_id)
+            comment_count = ArticleComment.objects.filter(
+                article=article).count()
+            reply_to = None
+
+            if 'reply_to' in data:
+                reply_to = ArticleComment.objects.get(
+                    pk=data['reply_to'])
+
             article_comment = ArticleComment(
-                comment = data['comment'],
-                user = user,
-                article = article,
-                parent_comment = parent_comment
+                comment=data['comment'],
+                user=user,
+                article=article,
+                reply_to=reply_to,
+                display_id=comment_count + 1
             )
             article_comment.save()
 
-            serializer = ArticleCommentSerializer(article_comment)
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request})
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         except (IntegrityError, ValueError, KeyError) as e:
             print(e)
@@ -67,8 +80,9 @@ def article_comment_view(request, article_id):
     # end if
 # def
 
+
 @api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes((IsAuthenticated,))
+@permission_classes((IsAuthenticatedOrReadOnly,))
 def single_article_comment_view(request, article_id, pk):
     '''
     Get an article comment by primary key/ id
@@ -76,13 +90,15 @@ def single_article_comment_view(request, article_id, pk):
     if request.method == 'GET':
         try:
             article_comment = ArticleComment.objects.get(pk=pk)
-            serializer = ArticleCommentSerializer(article_comment)
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request})
             return Response(serializer.data, status=status.HTTP_200_OK)
         except (ObjectDoesNotExist, KeyError, ValueError) as e:
             print(e)
             return Response(status=status.HTTP_400_BAD_REQUEST)
         # end try-except
     # end if
+
     '''
     Update comment
     '''
@@ -91,11 +107,16 @@ def single_article_comment_view(request, article_id, pk):
         try:
             article_comment = ArticleComment.objects.get(pk=pk)
 
+            if article_comment.user != request.user:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
+            # end if
+
             if 'comment' in data:
                 article_comment.comment = data['comment']
 
             article_comment.save()
-            serializer = ArticleCommentSerializer(article_comment)
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request})
             return Response(serializer.data, status=status.HTTP_200_OK)
         except ArticleComment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -103,12 +124,18 @@ def single_article_comment_view(request, article_id, pk):
             return Response(status=status.HTTP_400_BAD_REQUEST)
         # end try-except
     # end if
+
     '''
     Deletes an article comment
     '''
     if request.method == 'DELETE':
         try:
             article_comment = ArticleComment.objects.get(pk=pk)
+
+            if article_comment.user != request.user:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
+            # end if
+
             article_comment.delete()
             return Response(status=status.HTTP_200_OK)
         except ArticleComment.DoesNotExist:
@@ -116,3 +143,120 @@ def single_article_comment_view(request, article_id, pk):
         # end try-except
     # end if
 # def
+
+
+@api_view(['PATCH'])
+@permission_classes((IsAuthenticatedOrReadOnly,))
+def pin_comment_view(request, article_id, pk):
+    '''
+    Pins comment
+    '''
+    if request.method == 'PATCH':
+        data = request.data
+        try:
+            article_comment = ArticleComment.objects.get(pk=pk)
+
+            if article_comment.user != request.user:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
+            # end if
+
+            article_comment.pinned = True
+            article_comment.save()
+
+            article_comment.save()
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request})
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        except ArticleComment.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        except (KeyError, ValueError) as e:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        # end try-except
+    # end if
+# end def
+
+
+@api_view(['PATCH'])
+@permission_classes((IsAuthenticatedOrReadOnly,))
+def unpin_comment_view(request, article_id, pk):
+    '''
+    Unpins comment
+    '''
+    if request.method == 'PATCH':
+        data = request.data
+        try:
+            article_comment = ArticleComment.objects.get(pk=pk)
+
+            if article_comment.user != request.user:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
+            # end if
+
+            article_comment.pinned = False
+            article_comment.save()
+
+            article_comment.save()
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request})
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        except ArticleComment.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        except (KeyError, ValueError) as e:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        # end try-except
+    # end if
+# end def
+
+
+@api_view(['POST', 'DELETE'])
+@permission_classes((IsAuthenticatedOrReadOnly,))
+def article_comment_engagement_view(request, article_id, pk):
+    '''
+    Like a comment
+    '''
+    if request.method == 'POST':
+        user = request.user
+        try:
+            article_comment = ArticleComment.objects.get(pk=pk)
+
+            if ArticleCommentEngagement.objects.filter(comment=article_comment).filter(user=user).exists():
+                return Response(NestedArticleCommentSerializer(article_comment, context={'request': request, 'recursive': True}).data, status=status.HTTP_409_CONFLICT)
+            # end if
+
+            engagement = ArticleCommentEngagement(
+                comment=article_comment,
+                user=user
+            )
+            engagement.save()
+            article_comment.save()
+
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request, 'recursive': True})
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        except ObjectDoesNotExist as e:
+            print(e)
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        # end try-except
+    # end if
+
+    '''
+    Unlike a comment
+    '''
+    if request.method == 'DELETE':
+        user = request.user
+        try:
+            article_comment = ArticleComment.objects.get(pk=pk)
+
+            engagement = ArticleCommentEngagement.objects.filter(
+                comment=article_comment).get(user=user)
+            engagement.delete()
+            article_comment.save()
+
+            serializer = NestedArticleCommentSerializer(
+                article_comment, context={'request': request, 'recursive': True})
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        except ObjectDoesNotExist as e:
+            print(e)
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        # end try-except
+    # end if
+# end def
